Updated MerryFairyTalesAPI.yaml
The swagger page now reflects that the add and update book API calls will consume JSON objects rather than form data. I also removed some of the unnecessary API calls, such as those for the book pages. The calls that upload or download the entire book JSON object can be used to handle the adding, editing, or deleting of pages.